+++ rules/enum.go
