+++ rules/game.go
