+++ rules/position.go
